data-service/requirements.txt
fastapi[all]==0.68.0
python-jose==3.3.0
sqlalchemy==1.4.26
passlib==1.7.4

